# Conflicts:
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/app_views/RecipeEditorPage.kt
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/data/FoodComponent.kt
@@ -34,5 +34,5 @@ sealed interface FoodComponent {
     /**
      * Returns the servings of the food component
      */
-    val servings: Int
+    val servings: Double
 }
