+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/data/Recipe.kt
@@ -9,7 +9,7 @@ data class Recipe(
     override val carbohydrates: Double = 0.0,
     override val protein: Double = 0.0,
     override val fat: Double = 0.0,
-    override val servings: Int = 1,
+    override val servings: Double = 1.0,
     var ingredients: List<Ingredient> = emptyList(),
     val instructions: String = "",
     val visibility: RecipeVisibility = RecipeVisibility.OWNER
