+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/persistence/entity/IngredientEntity.kt
@@ -31,6 +31,6 @@ data class IngredientEntity (
     val recipeId: String,
     val foodProductId: String,
     val quantity: Double,
-    val servings: Int,
+    val servings: Double,
     val servingSize: ServingSize
 )
